requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
PyYAML>=5.3.1
matplotlib>=3.2.2
pandas>=1.1.4
seaborn>=0.11.0
tqdm>=4.41.0
pillow>=7.1.2
tensorboard>=2.4.1 